// task3/src/app/models/ContributorsModel.ts
export class ContributorsModel {
    id: String;
    login: String;
    avatar_url: String;
    html_url: String;
    followers_url: String;
    repos_url: String;
    contributions: number;
    followers: number;
    reposGists: number;
    repoArr: any[];
    repoArrDetails: String[];
    public_repos: number;
    public_gists: number;
    name: String;
}